remove unneeded search icon component
